Change ArgbToAlphaAndColorRef() to return uint8_t for the alpha value
The alpha value is always in the range [0, 255].

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/123890
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fxge/dib/fx_dib.cpp
@@ -53,7 +53,7 @@ FX_BGR_STRUCT<uint8_t> ArgbToBGRStruct(FX_ARGB argb) {
   return {FXARGB_B(argb), FXARGB_G(argb), FXARGB_R(argb)};
 }
 
-std::pair<int, FX_COLORREF> ArgbToAlphaAndColorRef(FX_ARGB argb) {
+std::pair<uint8_t, FX_COLORREF> ArgbToAlphaAndColorRef(FX_ARGB argb) {
   return {FXARGB_A(argb), ArgbToColorRef(argb)};
 }
